+++ git-odb/src/pack/tree/from_offsets.rs
@@ -47,7 +47,7 @@ impl<T> Tree<T> {
         );
 
         let anticpiated_num_objects = if let Some(num_objects) = data_sorted_by_offsets.size_hint().1 {
-            progress.init(Some(num_objects as u32), Some("objects"));
+            progress.init(Some(num_objects), Some("objects"));
             num_objects
         } else {
             0
